esm: move TreeIndex out of the bincode cache into .esm.tree (rkyv)
Stage 4 of the bincode -> rkyv migration (see todos.md / RUSTSEC-2025-0141).
This is the first production type wired through rkyv::access_unchecked --
every prior stage (rkyvcache.rs, the Index API reshape) was building toward
this, and its own test suite was validated only against a throwaway type
until now.

TreeIndex/GroupEntry/ChildRef gain rkyv Archive/Serialize/Deserialize derives
and lose their serde ones (neither is ever JSON-exported, and after this
change neither is bincode-encoded either). Their usize-typed arena indices
(GroupEntry::parent, ChildRex::Group, TreeIndex::roots, offset_map's values)
become u32 -- this crate's rkyv is pinned to 32-bit pointer width, and the
convention is to pin every archived field width explicitly rather than lean
on that pin as the only guarantee. usize stays at the in-memory API boundary
(TreeView's methods); conversion happens at the edge.

TreeView (introduced storage-agnostic in the prior stage) now wraps
Option<&Archived<TreeIndex>> instead of &TreeIndex -- Section::Absent is a
real state (lite mode, or no cache built yet), and every existing lib.rs
caller already expected a working empty-result answer in that case rather
than a panic; every TreeView method preserves that exactly, verified by a
dedicated absent-state test. A synthetic-ESM round-trip test with nested
GRUPs and both ChildRef variants proves the writer's archive and the
archived-field read idioms (endian-wrapped scalars needing .to_native(),
offset_map's key needing conversion to its archived type before lookup,
since rend's wrapper types have no Borrow<T> for the unwrapped type) agree
with the plain in-memory structure.

CacheFile (the remaining bincode blob: form_index, edid_index, xref_index,
search_index) loses its tree field, which is exactly the kind of byte-shape
change this crate's CACHE_VERSION convention exists for -- bumped 12 -> 13.
try_load_cache now requires both the bincode CacheFile and the .esm.tree
section to validate against the same ESM identity before considering the
cache usable; build_fresh writes the freshly-built tree to its own section
and re-maps it rather than keeping a second in-memory representation around,
so there is exactly one code path for how Index ever reads tree data.

Verified against the real 20260724 snapshot (5.8M records): a cold `tree`
build recreates .esm.tree correctly, a warm reload does not rewrite either
cache file (proving the two-piece validation path loads rather than
rebuilds), and `tree`/`get`/`list --type WEAP` all produce correct output.

+++ esm/src/lib.rs
@@ -1056,7 +1056,13 @@ impl Database {
         for child in children_slice {
             match child {
                 ChildRef::Group(idx) => {
-                    result.push(GroupChild::Group(tree.group_node(idx)));
+                    // `ChildRef::Group` stores its arena index as `u32` (Stage
+                    // 4 pinned every `TreeIndex`-adjacent stored index to
+                    // `u32` for portable rkyv layout — see `tree.rs`), while
+                    // `TreeView::group_node` keeps `usize` at the in-memory
+                    // API boundary. Lossless widening cast, not a narrowing
+                    // one.
+                    result.push(GroupChild::Group(tree.group_node(idx as usize)));
                 }
                 ChildRef::Record {
                     form_id,
